requirements.txt
flask
mysqlclient
sqlalchemy
flask-sqlalchemy
werkzeug
flask-login
wtforms
python-dotenv
flask-wtf
flask-migrate
flask-mail
pyjwt
flask-bootstrap
flask-moment
flask-babel
click
guess_language-spirit
requests
googletrans
lxml
cairosvg
tinycss
cssselect
pygal